+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/mania/widgets/ManiaWidgetPlayerRankController.java
@@ -156,6 +156,7 @@ public class ManiaWidgetPlayerRankController extends WidgetController implements
 
   @FXML
   private void onReload() {
+    maniaClient.getAccountClient().resetRankedAccountsCache();
     rankedPlayersAvatarCache.clear();
     refresh();
   }